+++ examples/flow-react-example/src/Object.gen.js
@@ -8,8 +8,10 @@
 // $FlowExpectedError: Reason checked type sufficiently
 import * as ObjectBS from './Object.bs';
 
-export type openObject<T1> = {| +foo: number |};
+export type openObject<T1> = { +foo: number };
 
 export const foo: (openObject<mixed>) => openObject<mixed> = ObjectBS.foo;
 
-export const useProps: ({| +foo: number |}) => number = ObjectBS.useProps;
+export const useProps: ({ +foo: number }) => number = ObjectBS.useProps;
+
+export const usePropsClosed: ({| +fooClosed: number |}) => number = ObjectBS.usePropsClosed;